refactor tokenLogoURLs
